refactor(runtime): expose public API via app.runtime package
Empty __init__.py forced every business module to drill into
app.runtime.<submodule>, contradicting the tutorial promise that
authors "needn't read app/runtime/*". Re-export the 12 symbols
business code uses (Data, Key, DedupKey, Version, AdminOnly, Source,
Stream, bind, emit, node, query, wire) and switch every domain /
node / wiring / bridge / deployment site + tutorial examples to the
new entry. Cookbook #4 swaps select_latest for the equivalent query()
so the "import only from app.runtime" promise holds end-to-end.

Engine internals (Runtime, compile_graph, registries, durable
plumbing, migrator, http_source) stay submodule-only by design;
#6 of the tutorial spells out that Runtime is the framework launcher,
not a node-author API.
