+++ apps/agent-service/app/domain/fragment.py
@@ -15,7 +15,7 @@ from __future__ import annotations
 
 from typing import Annotated
 
-from app.runtime.data import Data, Key
+from app.runtime import Data, Key
 
 
 class Fragment(Data):
