+++ apps/agent-service/app/domain/message_request.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from typing import Annotated
 
-from app.runtime.data import Data, Key
+from app.runtime import Data, Key
 
 
 class MessageRequest(Data):
